minor update, refactor code
