+++ yoloface_gpu.py
@@ -1,7 +1,7 @@
 import argparse
 
 from PIL import Image
-from YOLO import YOLO, detect_video
+from YOLO import YOLO, detect_video, detect_img
 
 
 #####################################################################
@@ -27,21 +27,6 @@ def get_args():
                         help='image/video output path')
     args = parser.parse_args()
     return args
-
-
-def detect_img(yolo):
-    while True:
-        img = input('[i] ==> Input image filename: ')
-        try:
-            image = Image.open(img)
-        except:
-            print('[!] ==> Open Error! Try again!')
-            continue
-        else:
-            res_image = yolo.detect_image(image)
-            res_image.show()
-
-    yolo.close_session()
 
 
 def _main():
